Add files via upload
This is a project I built in order to get more familiar with Pygame. Unfortunately, the only pygame features I needed were drawing a rectangle, opening a screen, and placing text on a screen. The rest of the project (so far) only required basic Python objects and functions. I am still learning coding, so I do plan on adding and updating functionality and visuals (using sprites) for this project as time passes. Thank you.